+++ mlzero/__init__.py
@@ -1,3 +1,6 @@
 from mlzero import deep_learning
 from mlzero import supervised_learning
 from mlzero import unsupervised_learning
+from mlzero import utils
+
+VERSION = "0.0.3"