+++ CHANGES.md
@@ -1,13 +1,17 @@
-# Changes list. 
+# Changes list
 
 Versions are following semantic versioning (https://semver.org/) 
 
+## Version 0.0.3 (03.04.2022)
+
+- Added Support Vector Classifier(SVC) with QP solver
+- Added K-Nearest Neighbors(KNN) Classifier
 
 ## Version 0.0.2 (04.02.2022)
 
 - Added Linear Regression with SGD
 - Added Lasso, Ridge, ElasticNet regressions
-- Added L1, L2, L1 + L2 regularizations
+- Added L1, L2, L1 + L2 regularization functions
 
 ## Version 0.0.1 (16.01.2022)